requirements.txt
# -----------------
# Core ML Frameworks
# -----------------
torch==2.1.0
torchvision==0.16.0

# PyTorch Lightning
lightning==2.1.0
pytorch-lightning==2.1.0

# Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Computer Vision
opencv-python==4.8.0.76
mediapipe==0.10.7

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# NLP / Autocorrect
textblob==0.17.1
pyspellchecker==0.7.2

# Backend API (optional)
fastapi==0.103.2
uvicorn==0.23.2
python-multipart==0.0.6

# Model Export
onnx==1.14.1
onnxruntime==1.15.1

# Utility
tqdm==4.66.1
python-dotenv==1.0.0

# Jupyter
jupyter==1.0.0
notebook==7.0.3

torchmetrics~=1.8.2
soupsieve~=2.8.1
beautifulsoup4~=4.14.3
h11~=0.16.0
defusedxml~=0.7.1
ipython~=8.18.1
pillow~=11.3.0
pip~=22.3.1
attrs~=25.4.0
wheel~=0.38.4
tornado~=6.5.4
jinja2~=3.1.6
pyzmq~=27.1.0
pexpect~=4.9.0
cffi~=2.0.0
setuptools~=65.5.1
comm~=0.2.3
fqdn~=1.5.1
parso~=0.8.5
jedi~=0.19.2
lark~=1.3.1
regex~=2025.11.3
nltk~=3.9.2
scipy~=1.13.1
click~=8.1.8
requests~=2.32.5
pyparsing~=3.0.9
pytz~=2025.2
packaging~=24.2
QtPy~=2.4.3
ipywidgets~=8.1.8
pyyaml~=6.0.3
idna~=3.11
multidict~=6.7.0
propcache~=0.4.1
zipp~=3.23.0
anyio~=3.7.1
sniffio~=1.3.1
arrow~=1.4.0
python-dateutil~=2.9.0.post0
babel~=2.17.0
pygments~=2.19.2
httpcore~=1.0.9
httpx~=0.28.1
json5~=0.12.1
overrides~=7.7.0
threadpoolctl~=3.6.0
mpmath~=1.3.0
sympy~=1.14.0
platformdirs~=4.4.0
decorator~=5.2.1
psutil~=7.1.3
urllib3~=2.6.2
fsspec~=2024.12.0
bleach~=6.2.0
webencodings~=0.5.1
tinycss2~=1.4.0
aiohttp~=3.13.2
yarl~=1.22.0
markupsafe~=3.0.3
joblib~=1.5.3
aiosignal~=1.4.0
frozenlist~=1.8.0
aiohappyeyeballs~=2.6.1
certifi~=2025.11.12
debugpy~=1.8.19
traitlets~=5.14.3
django~=4.2.27
starlette~=0.27.0
pydantic~=2.12.5
exceptiongroup~=1.3.1
wcwidth~=0.2.14
image~=1.5.33
ptyprocess~=0.7.0
cycler~=0.12.1
six~=1.17.0
nbformat~=5.10.4
nbclient~=0.10.2
fastjsonschema~=2.21.2
jsonschema~=4.25.1
networkx~=3.2.1
jupyterlab~=4.5.1
asttokens~=3.0.1
contourpy~=1.3.0
executing~=2.2.1
fonttools~=4.60.2
ipykernel~=6.31.0
websocket-client~=1.9.0
sounddevice~=0.5.3
flatbuffers~=25.12.19
nbconvert~=7.16.6
pandocfilters~=1.5.1
mistune~=3.2.0
pycparser~=2.23
qtconsole~=5.7.0
terminado~=0.18.1
referencing~=0.36.2
webcolors~=24.11.1
jsonpointer~=3.0.0
isoduration~=20.11.0
kiwisolver~=1.4.7
Send2Trash~=1.8.3
filelock~=3.19.1
tomli~=2.3.0
humanfriendly~=10.0
coloredlogs~=15.0.1